refactor: merge Analysis page into Insights, remove duplicate charts
